requirements.txt
django>=5.1,<5.2
gunicorn>=22,<23
dj-database-url>=2,<3
whitenoise[brotli]>=6,<7
requests~=2.32.3

# Uncomment these lines to use a Postgres database. Both are needed, since in production
# (which uses Linux) we want to install from source, so that security updates from the
# underlying Heroku stack image are picked up automatically, thanks to dynamic linking.
# On other platforms/in development, the precompiled binary package is used instead, to
# speed up installation and avoid errors from missing libraries/headers.
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"

djangorestframework~=3.15.2
dj-config-url~=0.1.1